+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Arista EOS Role for Ansible
 - adds trunk_groups argument to eos_switchport
 - changes trunk_allowed_vlans to allow a range of vlans
 - changes arguments used by eos_config (see documentation)
+- fixes an issue with out of order vlans in eos_switchport
 
 ### New Modules